Move dimensions to mocks too
